iscsi-long.d: always clear started[arg0] timestamp after command completion
Brendan noticed that the script neglected to do this, which means that
DTrace was slowly building up an array full of now-unneeded timestamps.
This is bad practice and uses more DTrace memory than necessary, which
could result in long-term problems. So now we clear the timestamp after
we're done with it.

Note: this fix is not quite completely correct, because we clear the
timestamp only when things go through state E3. I believe that a command
that errors out or aborts may skip E3 and go straight to some other
state, but this basically never happens in our environment and I'm not
sure what all of the states I need to catch are. Covering state E3 fixes
all or almost all of the problem in the short term.

(Also, our key is the iSCSI structure address and these probably get
recycled and reused regularly due to slab allocation and so on.)