Update with information about problems and Solaris version dependencies
This updates the README and several scripts to note some problems people
may run into when running the scripts. Specifically: if you haven't
made any NFS exports, the NFS v3 fbt tracepoints aren't available, and
if you're on something sufficiently more recent than Solaris 10 update
8 how the scripts try to go from a znode to the ZFS pool name (and in
general the pool's SPA) doesn't work because of a kernel structure
change. Both issues were found (and tested) by Raymond Scott of TI.

We can't automatically fix the znode-to-SPA issue, but the code now does
all of the dereferencing with a CPP macro instead of inline and I've
added a comment about the situation (and a version of the macro that
should work for later versions of Solaris; Scott is running S10U10 and
it worked for him).