zfs-mon.d: generate real info about pool ZIO sources
Our previous reports in the 'Pool ZIO' section were, shall we say, not
correct for Illumos (and I'm not certain that they'd ever been correct
for Solaris), through my lack of understanding of what the field I was
looking at actually meant. This has been corrected and we now print more
useful and more correct information.